+++ docs/CHANGELOG.md
@@ -6,5 +6,6 @@
 - Added retrieval of addresses, insurance, parcels, shipments, trackers, pickups, batches, webhooks, api_keys, users, and more
 - Added buy and refund shipment along with regenerating rates
 - Added Travis CI Shellcheck testing
-- Added documentation
+- Added complete user and developer documentation
 - Added executable build script for deployment
+- Added manual unit tests